# apps/api/openapi-v2.yml
openapi: 3.0.3
info:
  title: Edge-Cloud Orchestrator API
  description: Production API documentation for the Edge-Cloud Orchestrator Control Plane
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v2/api-keys/:
    get:
      summary: List my API keys
      tags:
        - api-keys
      responses:
        "200":
          description: Default Response
    post:
      summary: Create a new API key
      tags:
        - api-keys
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  pattern: ^[a-zA-Z0-9\s\-_]+$
                permissions:
                  type: array
                  items:
                    type: string
                expiresAt:
                  anyOf:
                    - anyOf:
                        - not: {}
                        - type: string
                          format: date-time
                    - type: "null"
              required:
                - name
              additionalProperties: false
        required: true
      responses:
        "200":
          description: Default Response
  /v2/api-keys/{id}:
    delete:
      summary: Revoke an API key
      tags:
        - api-keys
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
  /v2/carbon/policy:
    get:
      summary: Get active carbon optimization policy
      tags:
        - carbon
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  carbonWeight:
                    type: number
                  isActive:
                    type: boolean
                  activePolicy:
                    type: string
    patch:
      summary: Update carbon optimization weight
      tags:
        - carbon
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                carbonWeight:
                  type: number
                  minimum: 0
                  maximum: 1
              required:
                - carbonWeight
              additionalProperties: false
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  carbonWeight:
                    type: number
  /v2/carbon/report:
    get:
      summary: Get auditable tenant carbon report
      tags:
        - carbon
      parameters:
        - schema:
            type: string
            format: date-time
          in: query
          name: from
          required: false
        - schema:
            type: string
            format: date-time
          in: query
          name: to
          required: false
        - schema:
            type: string
            format: uuid
          in: query
          name: tenantId
          required: false
        - schema:
            type: string
            enum:
              - json
              - csv
            default: json
          in: query
          name: format
          required: false
      responses:
        "200":
          description: Default Response
  /v2/carbon/savings:
    get:
      summary: Get carbon savings metrics
      tags:
        - carbon
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 7
          in: query
          name: days
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalSavedGco2Today:
                    type: number
                  totalSavedGco2Week:
                    type: number
                  equivalentTreesPlanted:
                    type: number
                  savingsHistory:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                        savedGco2:
                          type: number
...